Variables for light/dark and made Quiz Card reusable by passing props
